Use git submodule for ccan (moving opt_*_floatval into miner.c)
