+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "libblkmaker"]
 	path = libblkmaker
 	url = git://gitorious.org/bitcoin/libblkmaker.git
+[submodule "ccan"]
+	path = ccan-upstream
+	url = git://git.ozlabs.org/~ccan/ccan
