+++ ccan.bfg/Makefile.am
@@ -1,0 +1,14 @@
+noinst_LIBRARIES	= libccan.a
+
+libccan_a_SOURCES = \
+	../ccan-upstream/ccan/build_assert/build_assert.h  \
+	../ccan-upstream/ccan/cast/cast.h  \
+	../ccan-upstream/ccan/compiler/compiler.h  \
+	../ccan-upstream/ccan/opt/helpers.c  \
+	../ccan-upstream/ccan/opt/opt.c  \
+	../ccan-upstream/ccan/opt/opt.h  \
+	../ccan-upstream/ccan/opt/parse.c  \
+	../ccan-upstream/ccan/opt/private.h  \
+	../ccan-upstream/ccan/opt/usage.c  \
+	../ccan-upstream/ccan/typesafe_cb/typesafe_cb.h
+libccan_a_CPPFLAGS = -I$(top_srcdir)/ccan
