--- ccan/Makefile.am
@@ -1,4 +1,0 @@
-noinst_LIBRARIES	= libccan.a
-
-libccan_a_SOURCES	= compiler/compiler.h opt/helpers.c opt/opt.c opt/opt.h opt/parse.c opt/private.h opt/usage.c typesafe_cb/typesafe_cb.h
-libccan_a_CPPFLAGS	= -I$(top_srcdir)
